fix:[WIP] setup streaming fetching on '/books/[id]/[page]' route to boost user experience
