--- src/app/books/[id]/[page]/@content/page.tsx
@@ -1,80 +1,0 @@
-import { Calendar, ChevronRight, Layers, Pencil } from 'lucide-react';
-import { Card, PageWrapper } from '~/components/atoms';
-import {
-  Breadcrumb,
-  SearchButton,
-  CollapsibleMenu,
-} from '~/components/molecules';
-import { SearchForm } from '~/components/molecules/search-form';
-import { MainContent } from '~/components/organisms';
-import { dummyCategories } from '~/lib/utils';
-
-export type DetailBookPageProps = {
-  params: {
-    id: string;
-    page: string;
-  };
-  searchParams: {
-    query: string;
-  };
-};
-
-const DetailBookPage = async (props: DetailBookPageProps) => {
-  return (
-    <PageWrapper className="flex">
-      <aside className="fixed flex h-screen w-3/12 flex-col gap-3 overflow-y-scroll border-r border-black/10 pb-28 pl-8 pr-5 pt-5">
-        <h2 className="text-2xl font-bold">Daftar Isi</h2>
-        {dummyCategories.map((category) => (
-          <CollapsibleMenu
-            variant="tableOfContent"
-            Icon={<ChevronRight id="book-marked" size={24} strokeWidth={3} />}
-            key={category.category}
-            isRootCategory
-            {...category}
-          />
-        ))}
-      </aside>
-      <div className="ml-auto flex w-9/12 flex-col gap-5 pl-5 pr-8 pt-5">
-        <div className="flex justify-between">
-          <Breadcrumb
-            paths={[
-              { title: 'Categories', link: 'categories' },
-              { title: 'Adab Murid Terhadap Diri Sendiri' },
-            ]}
-          />
-          <SearchButton
-            placeholder="Cari topik..."
-            size="medium"
-            className="gap-7 rounded-md"
-          >
-            <SearchForm />
-          </SearchButton>
-        </div>
-        <Card.Root className="flex w-full flex-col gap-1.5 bg-primary-light p-5 text-light-100">
-          <Card.Header className="p-0">
-            <Card.Title className="text-4xl">
-              Adab Murid Terhadap Diri Sendiri
-            </Card.Title>
-          </Card.Header>
-          <Card.Footer className="flex gap-7 p-0">
-            <div className="flex items-center gap-1">
-              <Pencil size={20} />
-              <small className="text-base">Abdul Somad Al Bukhari</small>
-            </div>
-            <div className="flex items-center gap-1">
-              <Layers size={20} />
-              <small className="text-base">124 halaman</small>
-            </div>
-            <div className="flex items-center gap-1">
-              <Calendar size={20} />
-              <small className="text-base">2020</small>
-            </div>
-          </Card.Footer>
-        </Card.Root>
-        <MainContent query={props?.searchParams?.query} />
-      </div>
-    </PageWrapper>
-  );
-};
-
-export default DetailBookPage;
