--- src/app/books/[id]/[page]/@search/default.tsx
@@ -1,3 +1,0 @@
-export default function Default() {
-  return <h1 className="text-9xl text-red-500">TAII</h1>;
-}
