--- src/app/books/[id]/[page]/layout.tsx
@@ -1,11 +1,0 @@
-export default function Layout(props: {
-  content: React.ReactNode;
-  search: React.ReactNode;
-}) {
-  return (
-    <>
-      {props.content}
-      {props.search}
-    </>
-  );
-}
