+++ src/app/books/[id]/[page]/page.tsx
@@ -1,3 +1,24 @@
+import { Calendar, ChevronRight, Layers, Pencil } from 'lucide-react';
+import { Suspense } from 'react';
+import { Card, PageWrapper } from '~/components/atoms';
+import { Breadcrumb, SearchButton } from '~/components/molecules';
+import { TableOfContentSkeleton } from '~/components/organisms/menu-outline';
+import { SearchForm } from '~/components/molecules/search-form';
+import { MainContent, MenuOutline, SearchTable } from '~/components/organisms';
+import { MainContentSkeleton } from '~/components/organisms/main-content';
+import { SearchTableSkeleton } from '~/components/organisms/search-table';
+import { Await, getCategories, getContents } from '~/lib/utils';
+
+export type DetailBookPageProps = {
+  params: {
+    id: string;
+    page: string;
+  };
+  searchParams?: {
+    query: string;
+  };
+};
+
 export async function generateStaticParams() {
   return [
     { id: 'dummy-template', page: '1' },
@@ -13,5 +34,77 @@ export async function generateStaticParams() {
   ];
 }
 
-// eslint-disable-next-line import/no-anonymous-default-export
-export default () => null;
+const DetailBookPage = async ({ searchParams }: DetailBookPageProps) => {
+  const contentPromise = getContents();
+  const categoriesPromise = getCategories();
+
+  return (
+    <PageWrapper className="flex">
+      <aside className="fixed flex h-screen w-3/12 flex-col gap-3 overflow-y-scroll border-r border-black/10 pb-28 pl-8 pr-5 pt-5">
+        <h2 className="text-2xl font-bold">Daftar Isi</h2>
+        <Suspense fallback={<TableOfContentSkeleton />}>
+          <Await promise={categoriesPromise}>
+            {({ categories }) => (
+              <MenuOutline
+                variant="tableOfContent"
+                outlines={categories}
+                Icon={<ChevronRight size={24} strokeWidth={3} />}
+                isRootCategory
+              />
+            )}
+          </Await>
+        </Suspense>
+      </aside>
+      <div className="ml-auto flex w-9/12 flex-col gap-5 pl-5 pr-8 pt-5">
+        <div className="flex justify-between">
+          <Breadcrumb
+            paths={[
+              { title: 'Categories', link: 'categories' },
+              { title: 'Adab Murid Terhadap Diri Sendiri' },
+            ]}
+          />
+          <SearchButton
+            placeholder="Cari topik..."
+            size="medium"
+            className="gap-7 rounded-md"
+          >
+            <SearchForm />
+          </SearchButton>
+        </div>
+        <Card.Root className="flex w-full flex-col gap-1.5 bg-primary-light p-5 text-light-100">
+          <Card.Header className="p-0">
+            <Card.Title className="text-4xl">
+              Adab Murid Terhadap Diri Sendiri
+            </Card.Title>
+          </Card.Header>
+          <Card.Footer className="flex gap-7 p-0">
+            <div className="flex items-center gap-1">
+              <Pencil size={20} />
+              <small className="text-base">Abdul Somad Al Bukhari</small>
+            </div>
+            <div className="flex items-center gap-1">
+              <Layers size={20} />
+              <small className="text-base">124 halaman</small>
+            </div>
+            <div className="flex items-center gap-1">
+              <Calendar size={20} />
+              <small className="text-base">2020</small>
+            </div>
+          </Card.Footer>
+        </Card.Root>
+        <Suspense fallback={<MainContentSkeleton />}>
+          <Await promise={contentPromise}>
+            {(content) => (
+              <MainContent query={searchParams?.query} content={content} />
+            )}
+          </Await>
+        </Suspense>
+      </div>
+      <Suspense fallback={<SearchTableSkeleton />}>
+        {searchParams?.query ? <SearchTable /> : null}
+      </Suspense>
+    </PageWrapper>
+  );
+};
+
+export default DetailBookPage;
