+++ src/components/organisms/search-table/index.ts
@@ -1,0 +1,3 @@
+import SearchTable, { SearchTableSkeleton } from './search-table.component';
+
+export { SearchTable, SearchTableSkeleton };
