+++ src/components/organisms/search-table/search-table.component.tsx
@@ -1,46 +1,9 @@
-import { Accordion, Table } from '~/components/atoms';
+import { Accordion, Skeleton, Table } from '~/components/atoms';
 import { Pagination } from '~/components/molecules';
+import { getInvoices } from '~/lib/utils';
 
-import { DetailBookPageProps } from '../@content/page';
-
-const invoices = [
-  {
-    no: 1,
-    page: 9,
-    relevantQuery: 'Hati Dari Sifat Tercela...',
-    title: 'Adab Murid Terhadap Diri Sendiri',
-  },
-  {
-    no: 12,
-    page: 19,
-    relevantQuery: 'hati, maka tidak dibenarkan menimbanya kecuali dala...',
-    title: 'Adab Murid Terhadap Diri Sendiri',
-  },
-  {
-    no: 13,
-    page: 19,
-    relevantQuery: 'hati (jantung).”...',
-    title: 'Adab Murid Terhadap Diri Sendiri',
-  },
-  {
-    no: 14,
-    page: 29,
-    relevantQuery: 'hatinya, menghiasi batinnya, mendekatkan diri kepad...',
-    title: 'Adab Murid Terhadap Diri Sendiri',
-  },
-  {
-    no: 151,
-    page: 139,
-    relevantQuery: 'sejatinya dia telah menukar sesuatu yang baik dengan ...',
-    title: 'Adab Murid Terhadap Diri Sendiri',
-  },
-];
-
-const SearchTable = async (props: DetailBookPageProps) => {
-  if (!props.searchParams.query) {
-    return null;
-  }
-
+const SearchTable = async () => {
+  const { invoices } = await getInvoices();
   // console.log(
   //   `/${props.params.id}/${props.params.page}?query=${props.searchParams.query}`,
   //   'SERVER'
@@ -89,5 +52,44 @@ const SearchTable = async (props: DetailBookPageProps) => {
     </Accordion.Root>
   );
 };
+
+export const SearchTableSkeleton = () => (
+  <div className="fixed bottom-0 right-0 w-9/12 bg-light-100 pb-3">
+    <p className="mx-auto bg-red-500 py-1 text-center text-xs text-white">
+      Tekan untuk menutup/membuka tabel pencarian
+    </p>
+    <Table.Root className="w-full table-fixed">
+      <Table.Caption>
+        <Pagination />
+      </Table.Caption>
+      <Table.Header className="w-full">
+        <Table.Row>
+          <Table.Head className="w-14">No.</Table.Head>
+          <Table.Head className="w-14">Hlm.</Table.Head>
+          <Table.Head className="w-2/3">Kueri Relevan</Table.Head>
+          <Table.Head className="w-1/3">Judul Buku</Table.Head>
+        </Table.Row>
+      </Table.Header>
+      <Table.Body>
+        {[...Array(5)].map((_, index) => (
+          <Table.Row key={index}>
+            <Table.Cell>
+              <Skeleton />
+            </Table.Cell>
+            <Table.Cell>
+              <Skeleton />
+            </Table.Cell>
+            <Table.Cell>
+              <Skeleton />
+            </Table.Cell>
+            <Table.Cell>
+              <Skeleton />
+            </Table.Cell>
+          </Table.Row>
+        ))}
+      </Table.Body>
+    </Table.Root>
+  </div>
+);
 
 export default SearchTable;
